+++ CHANGELOG.md
@@ -1,8 +1,13 @@
-# Changelog
+CHANGELOG.md# Changelog
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - No unreleased changes so far
+
+## [0.24.0] - In testing
+### Added
+- new topic /mqtt/topics that provides all known topics
+- new parameter --list-topics to print out new known topics to console
 
 ## [0.23.2] - 2022-01-10
 ### Fixed
@@ -408,7 +413,8 @@ Send empty message when topic is disabled
 ## [0.1.0] - 2021-05-27
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-mqtt/compare/v0.23.2...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-mqtt/compare/v0.24.0...HEAD
+[0.24.0]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.24.0
 [0.23.2]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.23.2
 [0.23.1]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.23.1
 [0.23.0]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.23.0